Fix `dup2_stdout` et al to not close the stdio file descriptor on error
The `dup2_std*` functions use the `?` operator to return errors from the
`dup2` system call. However, this means that the code after it
(`forget(target)`) is not run on errors, so `target` gets dropped closing
the file descriptor which violates I/O safety.

Fix that by using `ManuallyDrop` instead of `forget`.