Add io_uring flags available in linux-raw-sys (#1115)
* Add io_uring flags available in linux-raw-sys

* Make IORING_REGISTER_USE_REGISTERED_RING const

* Add io_uring_register_with and IoringRegisterFlags

IORING_REGISTER_USE_REGISTERED_RING must be ored into `opcode` when
calling `io_uring_register` with a registered ring fd. The value of this
flag is 1u32 << 31, and does not fit into the `IoringRegisterOp` enum.
To support this use case, add a new function `io_uring_register_with`
that takes a bitflags flags argument.

Also adds a minimal test that calls `io_uring_register` with a
registered ring fd. This operation isn't supported on older kernels, so
we ensure the feature is supported before completing the test.