Work around missing `TCGETS2`/`TCSETS2` on WSL.
WSL appears to be lacking support for `TCGETS2` and `TCSETS2`, so teach
rustix's `tcgetattr` and `tcsetattr` how to fall back to `TCGETS` and
`TCSETS` as needed.

This approach preserves rustix's ability to support arbitrary speed values,
while falling back as needed to support WSL.

This is expected to fix crossterm-rs/crossterm#912.