Fix `recv_uninit` to handle the `TRUNC` flag.
With the `TRUNC` flag, `recv` returns the untruncated length, which may
be longer than the provided buffer. Fix `recv_uninit` and
`recvfrom_uninit` to handle this case.

This mean that `recv_uninit` and `recvfrom_uninit` are unable to return
the full untruncated length returned by `recv` with the `TRUNC` flag,
however it appears that fixing that may require API changes.

Fixes #1153.