Don't enable the "fs" module implicitly.
Several configurations used to enable the "fs" module as rustix used to
use the "procfs" module for reading the AUXV vector on Linux. Rustix no
longer does this, as the "procfs" module's extra checks are incompatible
with QEMU on one hand, and not needed by most use cases on the other.

So since those configurations no longer use the "fs" module, we can
remove the code that enabled it.