Remove naked_functions feature usage for x86 (#722) (#1182)
* Remove naked_functions feature usage for x86 (#722)

Given that this function has no generics, it could easily use `global_asm!`
instead.

This is a backport of #722, except without the part that removes
outline asm for x86, as that's still needed on the 0.37 branch. It
fixes compilation errors on the 0.37 when compiled with the latest
nightly, due to `asm!` in `#[naked]` functions needed to be change
to `naked_asm!`.

Fixes #1181.

* Disable some tests and rebuilding of the archive files to fix building the 0.37 branch on current CI systems.

* Update Cirrus CI on the 0.37 branch

Remove freebsd-12, as it is no longer supported. And, update
freebsd-13 to freebsd-13-3.

* Don't assume that `target_env = "gnu"` implies Linux.

* Fix the cc crate to 1.0.68 for Rust 1.48 compat.

* Update CI to macos-12, actions/checkout@v4, node20.

---------

Co-authored-by: Gary Guo <gary@garyguo.net>