Fix tests on illumos.
Fix the layout of `epoll::Event` on x86 and x86_64 on illumos to match
libc's layout.

Adjust tests in tests/net/sockopt.rs to avoid using a buffer larger than
supported on illumos.

And change tests/pty/openpty.rs to avoid expecting we can write an EOF
to a pty on illumos, as that doesn't appear to work.