Remove procfs. (#1275)
Remove the procfs module, as most users don't need this functionality
and it can be implemented as a layer on top of rustix. This layer is now
available in the [rustix-linux-procfs] crate.

This requires changing `ttyname` to not use the `procfs` crate. It now
just uses a hard-coded "/proc/self/fd" path.

[rustix-linux-procfs]: https://crates.io/crates/rustix-linux-procfs