Support `SIGRTMIN+n` signals.
Change `Signal` from an `enum` to a `struct` wrapper around an `i32`
so that it can accomodate values it doesn't know about statically,
and add APIs for constructing and working with `SIGRTMIN+n` signal
values.

This requires making `Signal::from_raw` depend on a new "use-libc-sigrt"
feature, as it depends on being able to call into libc. This also adds
`Signal::from_raw_unchecked` which does not call into libc and is
unsafe.

Fixes #1249.