Clarify safety invariants in `SocketAddrArg`.
In making another safety review of the `SocketAddrArg` code, I found it
too subtle to have `MMsgHdr::new_with_addr` depending on knowing that
`with_msghdr` happens to avoid using a temporary in the case `SocketAddrAny`.
Change it to avoid doing this, add more comments, and change `with_msghdr`'s
argument to a reference, to prevent it from being accidentally used in
this way.

And change `with_noaddr_msghdr` to `noaddr_msghdr` and make it just return a
`msghdr`, because it never needs to construct any temporaries, so it
avoids this issue.

Also, make `SocketAddrArg::with_sockaddr` and `with_recv_msghdr` unsafe, as
the provided function will need to contain an `unsafe` block which will need
to be able to trust it's being called correctly.

And add safety comments for the callers and impls of these functions and
traits.