Rename `rustix_int_0x80` and protect it from multiple definitions. (#1396)
* Rename `rustix_int_0x80` and protect it from multiple definitions.

In lto builds, multiple versions of the `rustix_int_0x80` assembly code
can be included in the same .s file. To avoid conflicts when one of them
comes from an older version of rustix, rename the symbol to
`rustix_x86_int_0x80`, and to avoid conflicts with future versions, enclose
the code in an `ifndef` block.

Fixes #1394.

* Fix compilation of tests/event/port.rs

When libc/extra_trais is not enabled, `libc::timespec` does not have a
`PartialEq` implementation, so compare the fields directly.