Add a `Signal::from_named_raw` constructor.
Rust 1.0 moved `Signal::from_raw` into a [separate crate], since
properly supporting `SIGRTMIN+n` signals requires querying libc. But
it's still useful to have a version of this function in rustix which
is just documented to not support `SIGRTMIN+n` signals, so re-introduce
it, with a different name and with appropriatet documentation.

[separate crate]: https://crates.io/crates/rustix-libc-wrappers