DRAFT: Support `Buffer` on `&mut MaybeUninit<T>`
This returns an output of `Option<&mut T>`: `Some` for `.assume_init(1)`,
and `None` for `.assume_init(0)`.

This is a draft because it's not clear yet whether `Option` is the best
possible return value. For syscalls that will *always* initialize the
buffer if they return successfully, this isn't ideal, because it
requires handling an `Ok(None)` case that'll never happen. However, for
syscalls that *can* return zero output successfully, like `read`, we'd
need some way to handle that case.