Handle `Duration` overflow gracefully in `kevent`.
Add a `kevent_timespec` function, which is similar to `kevent`, but
which takes a `Timespec` rather than a `Duration`, so that it can behave
consistently with the other functions in `rustix::event`. This also
more closely reflects how the underlying `kevent` function takes a
`timespec` argument.

And, change the existing `kevent` to treat overflow in its `Duration`
argument by saturating to an effectively infinite timeout. This supports
users that use `Duration::MAX` to indicate an effectively infinite
timeout.