Expand the windows-sys dependency to include 0.60.0. (#1476)
* Expand the windows-sys dependency to include 0.60.0.

Expand the windows-sys dependency to ">=0.52, <0.61", to support the new
windows-sys 0.60.

windows-sys 0.60 does make one change that affected rustix, which is to
move the `BOOL` type from `windows_sys::Win32::Foundation::BOOL` to
`windows_sys::core::BOOL`, so to support both old and new versions, just
hard-code the `BOOL` type, as it's just an `i32` and always will be.

* Pin more packages for the MSRV build.

* Fix warnings.

* Disable more test on FreeBSD.