build: Use `OUT_DIR` for test compilation output (#1526)
The `can_compile` utility writes its temporary output file to the
system's global temporary directory. This can clutter the host system
with build artifacts.

By using Cargo's `OUT_DIR`, the temporary file is written to a
directory specific to the build, ensuring that build artifacts are
contained within the project's build output and do not mess up the
host system.

Follow-up to #1497.